refactor(backtracking): extract and relocate interactive demo flows to demos/backtracking
