+++ src/backtracking/graph_coloring.c
@@ -1,7 +1,6 @@
 #include "../utils/config.h"
 #include "backtracking.h"
 #include "cross_platform_timer.h"
-#include "safe_input.h"
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -12,15 +11,6 @@
 #ifdef _WIN32
 #include <windows.h>
 #endif
-
-#define MAX_V 10
-
-typedef struct
-{
-    const char* name;
-    int num_vertices;
-    int matrix[MAX_V][MAX_V];
-} GraphTopology;
 
 static const char* color_names[] = {"NONE", "RED", "GREEN", "BLUE", "YELLOW", "MAGENTA"};
 static const char* color_ansi[] = {
@@ -164,8 +154,8 @@ static void print_drawn_graph(const GraphTopology* graph, int colors[MAX_V], int
     printf("\n");
 }
 
-static void print_graph_state(const GraphTopology* graph, int colors[MAX_V], int current_vertex,
-                              const char* status_msg, int delay_time)
+void print_graph_state(const GraphTopology* graph, int colors[MAX_V], int current_vertex,
+                       const char* status_msg, int delay_time)
 {
     if (!is_instant())
     {
@@ -241,159 +231,43 @@ static int is_safe(int v, const GraphTopology* graph, int colors[MAX_V], int c)
     return 1;
 }
 
-static int solve_graph_coloring_util(const GraphTopology* graph, int m, int colors[MAX_V], int v,
-                                     int delay_time)
+int solve_graph_coloring_util(const GraphTopology* graph, int m, int colors[MAX_V], int v,
+                              int delay_time)
 {
-    // Base Case: All vertices are colored
+    // If all vertices are assigned colors, return true
     if (v == graph->num_vertices)
     {
-        print_graph_state(graph, colors, -1, "🎉 SUCCESS: All vertices colored successfully!",
-                          delay_time);
         return 1;
     }
 
     // Try different colors for vertex v
     for (int c = 1; c <= m; c++)
     {
-        char status_msg[128];
-        snprintf(status_msg, sizeof(status_msg),
-                 "Trying to assign color %s%s" ANSI_COLOR_RESET " to V%d", color_ansi[c],
-                 color_names[c], v);
-        print_graph_state(graph, colors, v, status_msg, delay_time);
-
+        // Check if assignment of color c to v is fine
         if (is_safe(v, graph, colors, c))
         {
             colors[v] = c;
-            snprintf(status_msg, sizeof(status_msg),
-                     "Safe! Assigned color %s%s" ANSI_COLOR_RESET " to V%d. Recursing...",
-                     color_ansi[c], color_names[c], v);
-            print_graph_state(graph, colors, v, status_msg, delay_time);
+            char msg[128];
+            snprintf(msg, sizeof(msg), "Assigned %s to Vertex V%d. Checking neighbors...",
+                     color_names[c], v);
+            print_graph_state(graph, colors, v, msg, delay_time);
 
+            // Recur to assign colors to rest of the vertices
             if (solve_graph_coloring_util(graph, m, colors, v + 1, delay_time))
             {
                 return 1;
             }
 
-            // Backtrack
+            // If assigning color c doesn't lead to a solution then remove it
             colors[v] = 0;
-            snprintf(status_msg, sizeof(status_msg),
-                     "Conflict down the line! Backtracking... unassigning V%d", v);
-            print_graph_state(graph, colors, v, status_msg, delay_time);
-        }
-        else
-        {
-            snprintf(status_msg, sizeof(status_msg),
-                     "Conflict! Cannot assign %s%s" ANSI_COLOR_RESET " to V%d (adjacent conflict)",
-                     color_ansi[c], color_names[c], v);
-            print_graph_state(graph, colors, v, status_msg, delay_time);
+            snprintf(msg, sizeof(msg), "Conflict detected! Backtracking color of Vertex V%d...", v);
+            print_graph_state(graph, colors, v, msg, delay_time);
         }
     }
 
     return 0;
 }
 
-void graph_coloring_demo(void)
-{
-    // Set console output to UTF-8 on Windows for Unicode support (●)
-#ifdef _WIN32
-    UINT old_cp = GetConsoleOutputCP();
-    SetConsoleOutputCP(CP_UTF8);
-#endif
-
-    // Predefined Graph Topologies
-    GraphTopology topologies[] = {
-        {"Cycle Graph C5",
-         5,
-         {{0, 1, 0, 0, 1}, {1, 0, 1, 0, 0}, {0, 1, 0, 1, 0}, {0, 0, 1, 0, 1}, {1, 0, 0, 1, 0}}},
-        {"Bipartite Graph K3,3",
-         6,
-         {{0, 0, 0, 1, 1, 1},
-          {0, 0, 0, 1, 1, 1},
-          {0, 0, 0, 1, 1, 1},
-          {1, 1, 1, 0, 0, 0},
-          {1, 1, 1, 0, 0, 0},
-          {1, 1, 1, 0, 0, 0}}},
-        {"Complete Graph K4", 4, {{0, 1, 1, 1}, {1, 0, 1, 1}, {1, 1, 0, 1}, {1, 1, 1, 0}}},
-        {"Complete Graph K5",
-         5,
-         {{0, 1, 1, 1, 1}, {1, 0, 1, 1, 1}, {1, 1, 0, 1, 1}, {1, 1, 1, 0, 1}, {1, 1, 1, 1, 0}}},
-        {"Wheel Graph W5",
-         5,
-         {{0, 1, 1, 1, 1}, {1, 0, 1, 0, 1}, {1, 1, 0, 1, 0}, {1, 0, 1, 0, 1}, {1, 1, 0, 1, 0}}},
-        {"Petersen Graph",
-         10,
-         {{0, 1, 0, 0, 1, 1, 0, 0, 0, 0},
-          {1, 0, 1, 0, 0, 0, 1, 0, 0, 0},
-          {0, 1, 0, 1, 0, 0, 0, 1, 0, 0},
-          {0, 0, 1, 0, 1, 0, 0, 0, 1, 0},
-          {1, 0, 0, 1, 0, 0, 0, 0, 0, 1},
-          {1, 0, 0, 0, 0, 0, 0, 1, 1, 0},
-          {0, 1, 0, 0, 0, 0, 0, 0, 1, 1},
-          {0, 0, 1, 0, 0, 1, 0, 0, 0, 1},
-          {0, 0, 0, 1, 0, 1, 1, 0, 0, 0},
-          {0, 0, 0, 0, 1, 0, 1, 1, 0, 0}}}};
-    int num_topologies = sizeof(topologies) / sizeof(topologies[0]);
-
-    while (1)
-    {
-        int graph_choice;
-        printf("\n--- Graph Coloring Topology ---\n");
-        for (int i = 0; i < num_topologies; i++)
-        {
-            printf("%d. %s (%d vertices)\n", i + 1, topologies[i].name, topologies[i].num_vertices);
-        }
-        int status = safe_input_int(&graph_choice, "Select topology (1-6), or -1 to exit: ", 1,
-                                    num_topologies);
-
-        if (status == INPUT_EXIT_SIGNAL)
-        {
-#ifdef _WIN32
-            SetConsoleOutputCP(old_cp); // restore code page
-#endif
-            printf("\nReturning to Backtracking menu...\n");
-            return;
-        }
-
-        if (status == 0)
-            continue;
-
-        const GraphTopology* selected_graph = &topologies[graph_choice - 1];
-
-        int M;
-        status = safe_input_int(
-            &M, "\nEnter the number of colors M (between 1 and 5), or -1 to exit: ", 1, 5);
-        if (status == INPUT_EXIT_SIGNAL)
-        {
-#ifdef _WIN32
-            SetConsoleOutputCP(old_cp);
-#endif
-            printf("\nReturning to Backtracking menu...\n");
-            return;
-        }
-        if (status == 0)
-            continue;
-
-        int delay_time = 150;
-        int colors[MAX_V] = {0};
-
-        print_graph_state(selected_graph, colors, 0, "Starting graph coloring...", delay_time);
-
-        clock_t start_t = clock();
-        int solved = solve_graph_coloring_util(selected_graph, M, colors, 0, delay_time);
-        clock_t end_t = clock();
-        double total_t = (double)(end_t - start_t) / CLOCKS_PER_SEC;
-
-        if (solved)
-        {
-            printf("\n🎉 SUCCESS: The graph was successfully colored using %d colors!\n", M);
-        }
-        else
-        {
-            printf("\n❌ FAILURE: No valid coloring exists for this graph with %d colors.\n", M);
-        }
-        printf("Time taken: %f seconds\n", total_t);
-    }
-}
 // --- TEST WRAPPER ---
 bool run_graph_coloring_test(int colors_allowed)
 {
