+++ src/backtracking/knights_tour.c
@@ -2,7 +2,6 @@
 #include "../utils/config.h"
 #include "backtracking.h"
 #include "cross_platform_timer.h"
-#include "safe_input.h"
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -33,7 +32,7 @@ static int compare_candidates(const void* a, const void* b)
     return ((MoveCandidate*)a)->degree - ((MoveCandidate*)b)->degree;
 }
 
-static void print_board(int N, int board[8][8], int knight_x, int knight_y)
+void print_knights_tour_board(int N, int board[8][8], int knight_x, int knight_y)
 {
     if (!is_instant())
     {
@@ -72,11 +71,11 @@ static void print_board(int N, int board[8][8], int knight_x, int knight_y)
 }
 
 // Standard Backtracking solver (no heuristic) - Good for small boards
-static int solve_standard(int x, int y, int move_count, int board[8][8], int N)
+int solve_standard(int x, int y, int move_count, int board[8][8], int N)
 {
     if (move_count == N * N)
     {
-        print_board(N, board, x, y);
+        print_knights_tour_board(N, board, x, y);
         return 1;
     }
 
@@ -88,14 +87,14 @@ static int solve_standard(int x, int y, int move_count, int board[8][8], int N)
         if (next_x >= 0 && next_x < N && next_y >= 0 && next_y < N && board[next_x][next_y] == -1)
         {
             board[next_x][next_y] = move_count;
-            print_board(N, board, next_x, next_y);
+            print_knights_tour_board(N, board, next_x, next_y);
 
             if (solve_standard(next_x, next_y, move_count + 1, board, N))
                 return 1;
 
             // Backtrack
             board[next_x][next_y] = -1;
-            print_board(N, board, x, y);
+            print_knights_tour_board(N, board, x, y);
         }
     }
     return 0;
@@ -116,11 +115,11 @@ static int get_degree(int x, int y, int board[8][8], int N)
 }
 
 // Warnsdorff Heuristic guided Backtracking solver - Fast for 8x8 boards
-static int solve_warnsdorff(int x, int y, int move_count, int board[8][8], int N)
+int solve_warnsdorff(int x, int y, int move_count, int board[8][8], int N)
 {
     if (move_count == N * N)
     {
-        print_board(N, board, x, y);
+        print_knights_tour_board(N, board, x, y);
         return 1;
     }
 
@@ -149,86 +148,18 @@ static int solve_warnsdorff(int x, int y, int move_count, int board[8][8], int N
         int next_y = y + y_move[move_idx];
 
         board[next_x][next_y] = move_count;
-        print_board(N, board, next_x, next_y);
+        print_knights_tour_board(N, board, next_x, next_y);
 
         if (solve_warnsdorff(next_x, next_y, move_count + 1, board, N))
             return 1;
 
         // Backtrack
         board[next_x][next_y] = -1;
-        print_board(N, board, x, y);
+        print_knights_tour_board(N, board, x, y);
     }
     return 0;
 }
 
-void knights_tour_demo(void)
-{
-    // Set console output to UTF-8 on Windows for Emojis
-#ifdef _WIN32
-    UINT old_cp = GetConsoleOutputCP();
-    SetConsoleOutputCP(CP_UTF8);
-#endif
-
-    int board[8][8];
-
-    while (1)
-    {
-        int N;
-        int status =
-            safe_input_int(&N, "\nEnter the board size N (between 4 and 8), or -1 to exit: ", 1, 8);
-
-        if (status == INPUT_EXIT_SIGNAL)
-        {
-#ifdef _WIN32
-            SetConsoleOutputCP(old_cp); // restore code page
-#endif
-            printf("\nReturning to Backtracking menu...\n");
-            return;
-        }
-
-        if (status == 0)
-            continue;
-
-        // Initialize board
-        for (int i = 0; i < N; i++)
-        {
-            for (int j = 0; j < N; j++)
-            {
-                board[i][j] = -1;
-            }
-        }
-
-        // Start from cell (0, 0)
-        board[0][0] = 0;
-
-        print_board(N, board, 0, 0);
-
-        clock_t start_t = clock();
-        int solved = 0;
-        // Use standard backtracking for smaller boards to show backtracking/failures,
-        // and Warnsdorff heuristic for larger boards (N >= 6) to solve instantly.
-        if (N <= 5)
-        {
-            solved = solve_standard(0, 0, 1, board, N);
-        }
-        else
-        {
-            solved = solve_warnsdorff(0, 0, 1, board, N);
-        }
-        clock_t end_t = clock();
-        double total_t = (double)(end_t - start_t) / CLOCKS_PER_SEC;
-
-        if (solved)
-        {
-            printf("\n SUCCESS: The knight completed the tour successfully!\n");
-        }
-        else
-        {
-            printf("\n❌ FAILURE: No tour exists for the knight from the starting position.\n");
-        }
-        printf("Time taken: %f seconds\n", total_t);
-    }
-}
 // --- TEST WRAPPER ---
 bool run_knights_tour_test(int n)
 {
